translator/pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Kabir <kabir@pynecone.io>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<4.0.0"
googletrans = "^3.1.0a0"
pynecone-io = ">=0.0.23"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]
toml = "^0.10.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
